src/requirements.txt
fastapi==0.63.0
itsdangerous==1.1.0
psycopg2-binary==2.8.6
pydantic==1.7.3
pytest==6.2.2
sendgrid==6.5.0
six==1.15.0
requests==2.25.1
Sphinx==3.4.3
sphinx-autobuild==2020.9.1
SQLAlchemy==1.3.22
starlette==0.13.6
uvicorn==0.13.3
passlib==1.7.4
PyJWT==1.1.0